♻️ Extract date logic into utility
