+++ static/app/views/alerts/rules/metric/details/utils.tsx
@@ -68,3 +68,59 @@ export function getFilter(rule: MetricRule): string[] | null {
   const eventType = extractEventTypeFilterFromRule(rule);
   return (query ? `(${eventType}) AND (${query.trim()})` : eventType).split(' ');
 }
+
+export function getViableDateRange({
+  interval,
+  rule: {timeWindow, dataset},
+  timePeriod: rawTimePeriod,
+}: {
+  interval: string;
+  rule: MetricRule;
+  timePeriod: TimePeriodType;
+}) {
+  const timePeriod = {...rawTimePeriod};
+
+  // Fix for 7 days * 1m interval being over the max number of results from events api
+  // 10k events is the current max
+  if (
+    timePeriod.usingPeriod &&
+    timePeriod.period === TimePeriod.SEVEN_DAYS &&
+    interval === '1m'
+  ) {
+    timePeriod.start = getUtcDateString(
+      // -5 minutes provides a small cushion for rounding up minutes. This might be able to be smaller
+      moment(moment.utc(timePeriod.end).subtract(10000 - 5, 'minutes'))
+    );
+  }
+
+  // If the chart duration isn't as long as the rollup duration the events-stats
+  // endpoint will return an invalid timeseriesData dataset
+  let viableStartDate = getUtcDateString(
+    moment.min(
+      moment.utc(timePeriod.start),
+      moment.utc(timePeriod.end).subtract(timeWindow, 'minutes')
+    )
+  );
+
+  // Events Analytics Platform Span queries only support up to 2016 buckets.
+  // 14 day 10m and 7 day 5m interval queries actually exceed this limit because we always extend the end date by an extra bucket.
+  // We push forward the start date by a bucket to counteract this and return to 2016 buckets.
+  if (
+    dataset === Dataset.EVENTS_ANALYTICS_PLATFORM &&
+    timePeriod.usingPeriod &&
+    ((timePeriod.period === TimePeriod.FOURTEEN_DAYS && interval === '10m') ||
+      (timePeriod.period === TimePeriod.SEVEN_DAYS && interval === '5m'))
+  ) {
+    viableStartDate = getUtcDateString(
+      moment.utc(viableStartDate).add(timeWindow, 'minutes')
+    );
+  }
+
+  const viableEndDate = getUtcDateString(
+    moment.utc(timePeriod.end).add(timeWindow, 'minutes')
+  );
+  return {
+    start: viableStartDate,
+    end: viableEndDate,
+  };
+}
